+++ custom_components/water_meter/config_flow.py
@@ -1,8 +1,8 @@
 """Config flow for Water Meter integration."""
 from homeassistant import config_entries
+from homeassistant.helpers import selector
 import voluptuous as vol
-from homeassistant.helpers import entity_registry as er
-from homeassistant.core import HomeAssistant
+
 from .const import DOMAIN, CONF_SOURCE_SENSOR, CONF_FRIENDLY_NAME
 
 
@@ -21,18 +21,12 @@ class WaterMeterConfigFlow(config_entries.ConfigFlow, domain=DOMAIN):
                 data=user_input,
             )
 
-        hass: HomeAssistant = self.hass
-        entity_registry = er.async_get(hass)
-        sensor_entities = [
-            entity.entity_id
-            for entity in entity_registry.entities.values()
-            if entity.entity_id.startswith("sensor.")
-        ]
-
         schema = vol.Schema(
             {
-                vol.Required(CONF_SOURCE_SENSOR): vol.In(sorted(sensor_entities)),
-                vol.Required(CONF_FRIENDLY_NAME): str
+                vol.Required(CONF_SOURCE_SENSOR): selector.EntitySelector(
+                    selector.EntitySelectorConfig(domain="sensor")
+                ),
+                vol.Required(CONF_FRIENDLY_NAME): str,
             }
         )